Refactor : connectPeerHandler 리팩토링
상세:
- 기존로직을 controller로 모두 옮기는 방향으로 리팩토링함
